fix: Exported enums in schema due to migrations error
